requirements.txt
altair==4.1.0
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
astropy==4.0.1.post1
attrs==20.1.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
boto3==1.14.47
botocore==1.17.47
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
confuse==1.3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
h2o==3.30.1.1
htmlmin==0.1.12
idna==2.10
ImageHash==4.1.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.1
missingno==0.4.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==6.1.3
numba==0.51.0
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pandas-profiling==2.8.0
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyarrow==1.0.1
pycparser==2.20
pydeck==0.4.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.2
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
streamlit==0.65.2
streamlit-pandas-profiling==0.0.4
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.10
validators==0.18.0
visions==0.4.4
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
